+++ src/parsing/pretty.ml
@@ -38,7 +38,6 @@ let _ = let open LpLexer in
     ; "end", END
     ; "fail", FAIL
     ; "flag", FLAG
-    ; "focus", FOCUS
     ; "generalize", GENERALIZE
     ; "have", HAVE
     ; "in", IN
@@ -267,7 +266,6 @@ let tactic : p_tactic pp = fun ppf { elt;  _ } ->
   | P_tac_assume ids ->
       out ppf "assume%a" (List.pp (pp_unit " " |+ param_id) "") ids
   | P_tac_fail -> out ppf "fail"
-  | P_tac_focus i -> out ppf "focus %i" i
   | P_tac_generalize id -> out ppf "generalize %a" ident id
   | P_tac_have (id, t) -> out ppf "have %a: %a" ident id term t
   | P_tac_induction -> out ppf "induction"