+++ src/parsing/syntax.ml
@@ -196,7 +196,6 @@ type p_tactic_aux =
   | P_tac_apply of p_term
   | P_tac_assume of p_ident option list
   | P_tac_fail
-  | P_tac_focus of int
   | P_tac_generalize of p_ident
   | P_tac_have of p_ident * p_term
   | P_tac_induction
@@ -367,7 +366,6 @@ let eq_p_tactic : p_tactic eq = fun {elt=t1;_} {elt=t2;_} ->
       b1 = b2 && Option.eq eq_p_rw_patt p1 p2 && eq_p_term t1 t2
   | P_tac_query q1, P_tac_query q2 -> eq_p_query q1 q2
   | P_tac_why3 so1, P_tac_why3 so2 -> so1 = so2
-  | P_tac_focus n1, P_tac_focus n2 -> n1 = n2
   | P_tac_simpl q1, P_tac_simpl q2 -> Option.eq eq_p_qident q1 q2
   | P_tac_generalize i1, P_tac_generalize i2 -> eq_p_ident i1 i2
   | P_tac_admit, P_tac_admit
@@ -564,7 +562,6 @@ let fold_idents : ('a -> p_qident -> 'a) -> 'a -> p_command list -> 'a =
     | P_tac_admit
     | P_tac_refl
     | P_tac_sym
-    | P_tac_focus _
     | P_tac_why3 _
     | P_tac_solve
     | P_tac_fail